refactor: ProductUiModel 변환 로직을 Mapper.kt로 이동
